Highlight when mouse hover, Highlight when drag body part to location, Increase Order in Layer when Dragging
Use combination of collision 2D and line renderer to achieve the "Highlight" function.
Check if drag the body parts to card location (On trigger Enter and if the overlapped object is being dragging) if so Highlight the Location card.
Increase the Order in Layers of all elements in a dragging object (hold it up), Decrease if release the mouse key (put it down).