soft Complete "Highlight" function, add Raycast for test
Highlight function: when dragging the needed body part, highlight white color, when dragging the card above the card location, highlight yellow color. When is not dragging the needed body part, do not highlight, but if the mouse is over the card location, highlight white.
Add Raycast for test: Debug Log all the names of the GameObject with ray casted collider2D. Test result: the Start button is sometimes not working because the collider2D is below the panel collider and could not be detected and triggered the OnMouseOver function.